feat: 1/4 master degree

+++ frontend-service/src/services/api.js
@@ -1,6 +1,10 @@
 import axios from "axios";
 import qs from "qs";
 
+const checkPermission = (role) => {
+    return role === "admin";
+};
+
 export const getAccessToken = async () => {
     let data = qs.stringify({
         'grant_type': 'client_credentials',
@@ -10,7 +14,7 @@ export const getAccessToken = async () => {
     let config = {
         method: 'post',
         maxBodyLength: Infinity,
-        url: 'https://172.28.80.1:9443/oauth2/token',
+        url: 'https://172.23.128.1:9443/oauth2/token',
         data: data
     };
     try {
@@ -53,14 +57,14 @@ export const getToken = async () => {
 
 // Cấu hình cơ bản cho axios
 const apiTeacherClient = axios.create({
-    baseURL: "https://172.28.80.1:8243/teacher/v1",
+    baseURL: "https://172.23.128.1:8243/teacher/v1",
     headers: {
         "Content-Type": "application/json",
     },
 });
 
 const apiLabClient = axios.create({
-    baseURL: "https://172.28.80.1:8243/lab/v1",
+    baseURL: "https://172.23.128.1:8243/lab/v1",
     headers: {
         "Content-Type": "application/json",
     },
@@ -77,6 +81,8 @@ apiTeacherClient.interceptors.request.use(async (config) => {
         throw error;
     }
     return config;
+}, (error) => {
+    return Promise.reject(error);
 });
 
 apiLabClient.interceptors.request.use(async (config) => {
@@ -90,7 +96,45 @@ apiLabClient.interceptors.request.use(async (config) => {
         throw error;
     }
     return config;
+}, (error) => {
+    return Promise.reject(error);
 });
+
+apiTeacherClient.interceptors.response.use(
+    (response) => response,
+    async (error) => {
+        if (error.response && error.response.status === 401) {
+            try {
+                const newToken = await getAccessToken();
+                if (newToken) {
+                    error.config.headers['Authorization'] = `Bearer ${newToken}`;
+                    return axios.request(error.config);
+                }
+            } catch (error) {
+                console.error("Lỗi khi lấy token mới:", error);
+            }
+        }
+        return Promise.reject(error);
+    }
+);
+
+apiLabClient.interceptors.response.use(
+    (response) => response,
+    async (error) => {
+        if (error.response && error.response.status === 401) {
+            try {
+                const newToken = await getAccessToken();
+                if (newToken) {
+                    error.config.headers['Authorization'] = `Bearer ${newToken}`;
+                    return axios.request(error.config);
+                }
+            } catch (error) {
+                console.error("Lỗi khi lấy token mới:", error);
+            }
+        }
+        return Promise.reject(error);
+    }
+);
 
 // Hàm lấy danh sách giáo viên
 export const getTeachers = async () => {
@@ -99,6 +143,7 @@ export const getTeachers = async () => {
             getTeachers {
                 id
                 name
+                account
                 departmentId {
                     id
                     name
@@ -149,6 +194,7 @@ export const getDepartments = async () => {
                 teachers {
                     id
                     name
+                    account
                 }
                 subjects {
                     id
@@ -181,42 +227,48 @@ export const getLabs = async () => {
 };
 
 // Hàm tạo mới một phòng thí nghiệm
-export const createLab = async (labData) => {
+export const createLab = async (userRole, labData) => {
+    if (!checkPermission(userRole)) {
+        return { error: "Bạn không có quyền thực hiện hành động này!" };
+    }
     try {
         const response = await apiLabClient.post("/labs", labData);
-        console.log("LONG", response.data);
-        return response.data;
+        return { data: response.data };
     } catch (error) {
         console.error("Lỗi khi gọi API createLab:", error);
-        throw error;
+        return { error: "Lỗi khi gọi API createLab. Vui lòng thử lại sau!" };
     }
 };
 
 // Hàm cập nhật thông tin phòng thí nghiệm
-export const updateLab = async (labId, updatedData) => {
+export const updateLab = async (userRole, labId, updatedData) => {
+    if (!checkPermission(userRole)) {
+        return { error: "Bạn không có quyền thực hiện hành động này!" };
+    }
     try {
         const response = await apiLabClient.put(`/labs/${labId}`, updatedData);
-        console.log(response.data);
-        return response.data;
+        return { data: response.data };
     } catch (error) {
-        console.error(`Lỗi khi gọi API updateLab với labId ${labId}:`, error);
-        throw error;
+        console.error(`Lỗi khi gọi API updateLab:`, error);
+        throw { error: "Lỗi khi gọi API updateLab. Vui lòng thử lại sau!" };
     }
 };
 
 // Hàm xóa phòng thí nghiệm
-export const deleteLab = async (labId) => {
+export const deleteLab = async (userRole, labId) => {
+    if (!checkPermission(userRole)) {
+        return { error: "Bạn không có quyền thực hiện hành động này!" };
+    }
     try {
         const response = await apiLabClient.delete(`/labs/${labId}`);
-        console.log(response.data);
-        return response.data;
+        return { data: response.data };
     } catch (error) {
-        console.error(`Lỗi khi gọi API deleteLab với labId ${labId}:`, error);
-        throw error;
+        console.error(`Lỗi khi gọi API deleteLab:`, error);
+        throw { error: "Lỗi khi gọi API deleteLab. Vui lòng thử lại sau!" };
     }
 };
 
-// Hàm lấy danh sách thiết bị trong phòng thí nghiệm theo labId
+// Hàm lấy danh sách thiết bị trongg thí nghiệm theo labId
 export const getDevicesByLab = async (labId) => {
     try {
         console.log(labId)
@@ -224,42 +276,50 @@ export const getDevicesByLab = async (labId) => {
         console.log(response.data);
         return response.data;
     } catch (error) {
-        console.error(`Lỗi khi gọi API getDevicesByLab với labId ${labId}:`, error);
+        console.error(`Lỗi khi gọi API getDevicesByLab:`, error);
         throw error;
     }
 };
 
 // Hàm tạo mới một phòng thí nghiệm
-export const createDevice = async (labId, deviceData) => {
+export const createDevice = async (userRole, labId, deviceData) => {
+    if (!checkPermission(userRole)) {
+        return { error: "Bạn không có quyền thực hiện hành động này!" };
+    }
     try {
         const response = await apiLabClient.post(`/labs/${labId}/devices`, deviceData);
-        console.log(response.data);
-        return response.data;
+        return { data: response.data };
     } catch (error) {
-        console.error("Lỗi khi gọi API createDevice:", error);
-        throw error;
+        console.error('Lỗi khi gọi API createDevice:', error);
+        throw { error: "Lỗi khi gọi API createDevice. Vui lòng thử lại sau!" };
     }
 };
 
 // Hàm cập nhật thiết bị trong phòng thí nghiệm
-export const updateDevice = async (labId, deviceId, deviceData) => {
+export const updateDevice = async (userRole, labId, deviceId, deviceData) => {
+    if (!checkPermission(userRole)) {
+        return { error: "Bạn không có quyền thực hiện hành động này!" };
+    }
     try {
         const response = await apiLabClient.put(`/labs/${labId}/devices/${deviceId}`, deviceData);
-        return response.data;
+        return { data: response.data };
     } catch (error) {
-        console.error(`Lỗi khi cập nhật thiết bị ${deviceId}:`, error);
-        throw error;
+        console.error(`Lỗi khi gọi API updateDevice:`, error);
+        throw { error: "Lỗi khi gọi API updateDevice. Vui lòng thử lại sau!" };
     }
 };
 
 // Hàm cập nhật thiết bị trong phòng thí nghiệm
-export const deleteDevice = async (labId, deviceId) => {
+export const deleteDevice = async (userRole, labId, deviceId) => {
+    if (!checkPermission(userRole)) {
+        return { error: "Bạn không có quyền thực hiện hành động này!" };
+    }
     try {
         const response = await apiLabClient.delete(`/labs/${labId}/devices/${deviceId}`);
-        return response.data;
+        return { data: response.data };
     } catch (error) {
-        console.error(`Lỗi khi xóa thiết bị ${deviceId}:`, error);
-        throw error;
+        console.error(`Lỗi khi gọi API deleteDevice:`, error);
+        throw { error: "Lỗi khi gọi API deleteDevice. Vui lòng thử lại sau!" };
     }
 };
 
@@ -293,12 +353,27 @@ export const createSchedule = async (labId, scheduleData) => {
 };
 
 // Hàm cập nhật thiết bị trong phòng thí nghiệm
-export const deleteSchedule = async (labId, scheduleId) => {
-    try {
+export const deleteSchedule = async (role, user, reserved, labId, scheduleId) => {
+    console.log(role, user, reserved, labId, scheduleId)
+    if (role === 'admin') {
+        // Nếu là admin thì cho phép xóa ngay
         const response = await apiLabClient.delete(`/labs/${labId}/schedule/${scheduleId}`);
-        return response.data;
-    } catch (error) {
-        console.error(`Lỗi khi xóa lịch  ${scheduleId}:`, error);
-        throw error;
+        return { data: response.data };
+    } else if (role === 'user') {
+        // Nếu là user, cần kiểm tra account của teacher
+        const teachers = await getTeachers(); // Lấy danh sách teachers
+        const teacher = teachers.find(t => t.id === reserved); // Tìm teacher có id là reserved
+        console.log(teacher)
+        if (teacher && teacher.account === user) {
+            // Nếu teacher có account trùng với user, cho phép xóa
+            const response = await apiLabClient.delete(`/labs/${labId}/schedule/${scheduleId}`);
+            return { data: response.data };
+        } else {
+            // Nếu không có quyền xóa
+            return { error: "Bạn không có quyền thực hiện hành động này!" };
+        }
+    } else {
+        // Nếu role không hợp lệ
+        return { error: "Role không hợp lệ!" };
     }
 };
